Move definition of class SourceInfo to SourceInfo.h
