+++ XcodeMLtoCXX/src/CodeBuilder.cpp
@@ -13,6 +13,7 @@
 #include "XcodeMlType.h"
 #include "XcodeMlEnvironment.h"
 #include "TypeAnalyzer.h"
+#include "SourceInfo.h"
 #include "CodeBuilder.h"
 #include "LibXMLUtil.h"
